requirements.txt
pyspark==3.5.0
pyyaml==6.0.1
pandas==2.0.3
numpy==1.24.3